+++ src/CoreEx/Events/Subscribing/ErrorHandling.cs
@@ -10,24 +10,37 @@ namespace CoreEx.Events.Subscribing
     public enum ErrorHandling
     {
         /// <summary>
-        /// Indicates that when the corresponding <i>error</i> occurs the underlying <see cref="System.Exception"/> will continue to bubble up the stack unchanged (unhandled).
+        /// Indicates that when the corresponding <i>error</i> occurs the underlying <see cref="System.Exception"/> will continue to bubble up the stack as unhandled.
         /// </summary>
-        /// <remarks>Results in a <see cref="EventSubscriberException"/> where the <see cref="Abstractions.IExtendedException.IsTransient"/> property is set (overridden) to <c>false</c>.</remarks>
         None,
 
         /// <summary>
-        /// Indicates that when the corresponding <i>error</i> occurs the underlying <see cref="System.Exception"/> will be rethrown as a <see cref="EventSubscriberException"/> effectively terminating the current subscriber.
+        /// The <i>ThrowSubscriberException</i> has been renamed to <see cref="Handle"/>; please amend to <see cref="Handle"/> as the <i>ThrowSubscriberException</i> has now been obsoleted and will soon be removed.
+        /// </summary>
+        /// <remarks>The <i>ThrowSubscriberException</i> name was confusing being the primary reason for being renamed.</remarks>
+        [System.Obsolete("The 'ThrowSubscriberException' has been renamed to 'Handle'; please amend to 'Handle' as the 'ThrowSubscriberException' has now been obsoleted and will soon be removed.", error: true)]
+        ThrowSubscriberException,
+
+        /// <summary>
+        /// The <i>TransientRetry</i> has been renamed to <see cref="Retry"/>; please amend to <see cref="Retry"/> as the <i>TransientRetry</i> has now been obsoleted and will soon be removed.
+        /// </summary>
+        /// <remarks>The <i>TransientRetry</i> name was confusing as transient is implied and the handling is simply to retry where possible.</remarks>
+        [System.Obsolete("The 'TransientRetry' has been renamed to 'Retry'; please amend to 'Retry' as the 'TransientRetry' has now been obsoleted and will soon be removed.", error: true)]
+        TransientRetry,
+
+        /// <summary>
+        /// Indicates that when the corresponding <i>error</i> occurs that it should be handled by the subscriber.
         /// </summary>
         /// <remarks>Results in a <see cref="EventSubscriberException"/> where the <see cref="Abstractions.IExtendedException.IsTransient"/> property is set (overridden) to <c>false</c>.
-        /// <para>Depending on the underlying messaging subsystem this may result in the likes of the message being deadlettered where supported.</para></remarks>
-        ThrowSubscriberException,
+        /// <para>Depending on the underlying messaging subsystem this may result in the likes of the message being deadlettered (or equivalent) where supported.</para></remarks>
+        Handle,
 
         /// <summary>
         /// Indicates that when the corresponding <i>error</i> occurs that it may be transient and should be retried (where possible).
         /// </summary>
         /// <remarks>Results in a <see cref="EventSubscriberException"/> where the <see cref="Abstractions.IExtendedException.IsTransient"/> property is set (overridden) to <c>true</c>.
         /// <para>A <i>retry</i> will <i>only</i> occur where the <see cref="EventSubscriberBase"/> implementation supports/enables.</para></remarks>
-        TransientRetry,
+        Retry,
 
         /// <summary>
         /// Indicates that when the corresponding <i>error</i> occurs this is expected and the current event/message should be completed without further processing and logging (i.e. silently).